Button and TextView
Made an empty TextView and empty Button to hold the information
regarding the result of the roll and result of the button.  These are
thereafter being used by their Id by using the findViewById. The Button
and the TextView has also been altered in the xml code, so they are
dependent on each others position